# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Omit_Tender.id-or-createdOn-or-updatedBy-or-updatedOn_": {
				"$ref": "#/components/schemas/Pick_Tender.Exclude_keyofTender.id-or-createdOn-or-updatedBy-or-updatedOn__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"CreateTenderProperties": {
				"$ref": "#/components/schemas/Omit_Tender.id-or-createdOn-or-updatedBy-or-updatedOn_"
			},
			"Partial_TenderItem_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"unit": {
						"type": "string"
					},
					"materialNetUnitAmount": {
						"type": "number",
						"format": "double"
					},
					"materialNetAmount": {
						"type": "number",
						"format": "double"
					},
					"materialActualNetAmount": {
						"type": "number",
						"format": "double"
					},
					"totalMaterialAmount": {
						"type": "number",
						"format": "double"
					},
					"feeNetUnitAmount": {
						"type": "number",
						"format": "double"
					},
					"feeNetAmount": {
						"type": "number",
						"format": "double"
					},
					"feeActualNetAmount": {
						"type": "number",
						"format": "double"
					},
					"totalFeeAmount": {
						"type": "number",
						"format": "double"
					},
					"tenderId": {
						"type": "number",
						"format": "double"
					},
					"num": {
						"type": "number",
						"format": "double"
					},
					"createdOn": {
						"type": "string",
						"format": "date-time"
					},
					"updatedOn": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "number",
						"format": "double"
					},
					"updatedBy": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"CreateTenderItemProperties": {
				"$ref": "#/components/schemas/Omit_TenderItem.id-or-num-or-createdOn-or-updatedBy-or-updatedOn_"
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Vilkrig x Alpinnovacio CRM API",
		"version": "1.0.0",
		"license": {
			"name": "UNLICENSED"
		},
		"contact": {}
	},
	"paths": {
...
		"/tenders": {
			"post": {
				"operationId": "CreateTender",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Partial_Tender_"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Creates a new tender with the provided properties. This operation is secured and\nrequires a JWT token with \"Tender:Create\" permission.",
				"tags": [
					"Tender"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Tender:Create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The properties required to create a new tender.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTenderProperties",
								"description": "The properties required to create a new tender."
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetTenders",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Partial_Tender_"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves a list of all tenders within the system. This endpoint requires\nauthentication and is protected by JWT tokens with the \"Tender:List\" permission.",
				"tags": [
					"Tender"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Tender:List"
						]
					}
				],
				"parameters": []
			},
			"delete": {
				"operationId": "DeleteTenders",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Deletes multiple tenders based on the provided array of IDs. This operation requires\nauthentication and \"Tender:Delete\" permission. The deletion affects only the tenders\nassociated with the authenticated user's tenant.",
				"tags": [
					"Tender"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Tender:Delete"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "An object containing an array of tender IDs to delete.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"ids": {
										"items": {
											"type": "number",
											"format": "double"
										},
										"type": "array"
									}
								},
								"required": [
									"ids"
								],
								"type": "object",
								"description": "An object containing an array of tender IDs to delete."
							}
						}
					}
				}
			}
		},
		"/tenders/{id}/copy": {
			"post": {
				"operationId": "CopyTender",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Partial_Tender_"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Copies an existing tender based on its ID. This operation is secured and\nrequires a JWT token with \"Tender:Create\" permission.",
				"tags": [
					"Tender"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Tender:Create"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the tender to be copied.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tenders/{id}/send": {
			"post": {
				"operationId": "SendEmail",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Sends an email with details about a specific tender. Secured with a JWT token and requires \"Tender:Create\" permission.\nThis method is typically used for notifying stakeholders about tender updates or invitations to tender.",
				"tags": [
					"Tender"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Tender:Create"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the tender to send.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Contains the message to be included in the email.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"message": {
										"type": "string"
									}
								},
								"required": [
									"message"
								],
								"type": "object",
								"description": "Contains the message to be included in the email."
							}
						}
					}
				}
			}
		},
		"/tenders/{id}": {
			"get": {
				"operationId": "GetTender",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Partial_Tender_"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Retrieves detailed information about a specific tender by their ID.\nThis endpoint is protected by JWT authentication, requiring \"Tender:Get\" permission.",
				"tags": [
					"Tender"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Tender:Get"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the tender to retrieve.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateTender",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Partial_Tender_"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Updates the information of an existing tender, identified by their ID. Access to this\nendpoint is restricted to authenticated requests with \"Tender:Update\" permission.",
				"tags": [
					"Tender"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant:Update"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the tender to update.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "A partial tender object containing the fields to update.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Tender_",
								"description": "A partial tender object containing the fields to update."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTender",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Deletes a single tender identified by their ID. This endpoint is secured with JWT authentication\nand requires \"Tender:Delete\" permission. The operation is restricted to tenders within the\nauthenticated user's tenant.",
				"tags": [
					"Tender"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Tender:Delete"
						]
					}
				],
				"parameters": [
					{
						"description": "The ID of the tender to delete.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tenders/{id}/items": {
			"get": {
				"operationId": "GetTenderItems",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Partial_TenderItem_"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves detailed information about a specific tender by their ID.\nThis endpoint is protected by JWT authentication, requiring \"Tender:Get\" permission.",
				"tags": [
					"Tender"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Tender:Get"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the tender to retrieve.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "AddTenderItem",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Partial_TenderItem_"
								}
							}
						}
					}
				},
				"description": "Adds a new item to an existing tender identified by its ID. This operation is secured with JWT and requires \"Tender:Get\" permission.\nUseful for modifying tender details dynamically during the tender process.",
				"tags": [
					"Tender"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Tender:Get"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the tender to which the item will be added.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "The properties required to create the new tender item.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTenderItemProperties",
								"description": "The properties required to create the new tender item."
							}
						}
					}
				}
			}
		},
		"/tenders/{id}/items-copy/{targetTenderId}": {
			"put": {
				"operationId": "CopyTenderItem",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Copies an existing tender item to another tender. This endpoint is protected by JWT authentication,\nrequiring \"Tender:Update\" permission. It allows users to duplicate tender items efficiently between tenders.",
				"tags": [
					"Tender"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Tender:Update"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the tender containing the item to copy.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "The ID of the target tender where the item will be copied.",
						"in": "path",
						"name": "targetTenderId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tenders/{id}/items/{itemId}": {
			"put": {
				"operationId": "UpdateTenderItem",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Partial_TenderItem-or-null_"
								}
							}
						}
					}
				},
				"description": "Updates a specific item within a tender by its item ID. This endpoint requires JWT authentication\nand is protected by \"Tender:Update\" permission, ensuring that only authorized modifications are allowed.",
				"tags": [
					"Tender"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Tender:Update"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the tender containing the item.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "The unique identifier of the item to update.",
						"in": "path",
						"name": "itemId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "A partial tender item object containing the fields to update.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_TenderItem_",
								"description": "A partial tender item object containing the fields to update."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "RemoveTenderItem",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Removes a specific item from a tender by its item ID. This operation requires JWT authentication\nand is secured with \"Tender:Update\" permission, designed to allow modifications only by authorized personnel.",
				"tags": [
					"Tender"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Tender:Update"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the tender from which the item will be removed.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "The unique identifier of the item to be removed.",
						"in": "path",
						"name": "itemId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/tenders/{id}/items/{itemId}/order": {
			"put": {
				"operationId": "UpdateTenderItemOrder",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Tender"
				],
				"security": [
					{
						"jwtToken": [
							"Tenant",
							"Tender:Update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "path",
						"name": "itemId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"side": {
										"type": "string",
										"enum": [
											"up",
											"down"
										]
									}
								},
								"required": [
									"side"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}